fix(build): move timeoutSignal to a crypto-free core/timeout module
The first cut added timeoutSignal to core/utils.ts and imported it from
the price/market modules. core/utils.ts carries a node:crypto require()
fallback (randomHex/randomUUID), and the new imports dragged that module
into the platform:'browser' bundles (impl/browser, impl/shared/wallet-api),
where esbuild cannot resolve the "crypto" builtin — breaking the CI build.
The package has top-level side effects (the Buffer polyfill in
impl/browser/index), so it cannot be marked sideEffects:false to tree-shake
the fallback away.

Move timeoutSignal into its own dependency-free core/timeout.ts and import
from there. Verified by bundling all 11 tsup entries (every platform) to a
temp outDir: no "Could not resolve crypto" error. No behavior change.
